2016-4-11 15:25 完成了AST中表达式的检测操作

+++ json/UnUCompiler/AST/ASTFloatNode.h
@@ -10,7 +10,7 @@
 
 namespace UnUCompiler
 {
-	class ASTFloatNode : public ASTValueNode<float>
+	class ASTFloatNode : public ASTValueNode<double>
 	{
 		friend class ASTNodeCreater;
 	public:
